+++ morpheus-graphql-core/src/Data/Morpheus/Types/Internal/AST/Base.hs
@@ -15,7 +15,7 @@ module Data.Morpheus.Types.Internal.AST.Base
   ( Ref (..),
     Position (..),
     Message (..),
-    FieldName (..),
+    FieldName,
     Description,
     OperationType (..),
     QUERY,
@@ -30,7 +30,7 @@ module Data.Morpheus.Types.Internal.AST.Base
     GQLErrors,
     TRUE,
     FALSE,
-    TypeName (..),
+    TypeName,
     Msg (..),
     intercalateName,
     toFieldName,
@@ -50,16 +50,23 @@ where
 
 import Data.Aeson
   ( FromJSON,
-    ToJSON,
+    Options (..),
+    ToJSON (..),
     Value,
+    defaultOptions,
     encode,
+    genericToJSON,
   )
 import Data.ByteString.Lazy (ByteString)
 import Data.Char (toLower)
 import Data.Morpheus.Rendering.RenderGQL
   ( RenderGQL (..),
-    fromText,
     renderGQL,
+  )
+import Data.Morpheus.Types.Internal.AST.Name
+  ( FieldName,
+    Name (..),
+    TypeName,
   )
 import Data.Text (intercalate, pack)
 import qualified Data.Text as T
@@ -130,15 +137,8 @@ withPosition pos (ValidationError m ps) = ValidationError m (ps <> maybeToList p
 type ValidationErrors = [ValidationError]
 
 toGQLError :: ValidationError -> GQLError
-toGQLError (ValidationError m p) = GQLError m p
+toGQLError (ValidationError m p) = GQLError m p Nothing
 {-# INLINE toGQLError #-}
-
--- instance Lift InternalError where
---   lift = liftString . readInternalError
-
--- #if MIN_VERSION_template_haskell(2,16,0)
---   liftTyped = liftTypedString . readInternalError
--- #endif
 
 msgInternal :: (Msg a) => a -> InternalError
 msgInternal = InternalError . readMessage . msg
@@ -152,6 +152,9 @@ class Msg a where
   msg :: a -> Message
   msgSepBy :: Text -> [a] -> Message
   msgSepBy t = Message . intercalate t . fmap (readMessage . msg)
+
+instance Msg TypeName where
+  msg Name {unName} = Message $ "\"" <> unName <> "\""
 
 instance Msg Message where
   msg = id
@@ -171,65 +174,16 @@ instance Msg Text where
 instance Msg Value where
   msg = msg . encode
 
--- FieldName : lower case names
-newtype FieldName = FieldName {readName :: Text}
-  deriving
-    (Generic)
-  deriving newtype
-    ( Show,
-      Ord,
-      Eq,
-      IsString,
-      ToString,
-      Hashable,
-      Semigroup,
-      FromJSON,
-      ToJSON
-    )
-
-instance Lift FieldName where
-  lift = liftString . readName
-
-#if MIN_VERSION_template_haskell(2,16,0)
-  liftTyped = liftTypedString . readName
-#endif
-
 instance Msg FieldName where
-  msg FieldName {readName} = Message $ "\"" <> readName <> "\""
-
-instance RenderGQL FieldName where
-  renderGQL = fromText . readName
+  msg Name {unName} = Message $ "\"" <> unName <> "\""
 
 intercalateName :: FieldName -> [FieldName] -> FieldName
-intercalateName (FieldName x) = FieldName . intercalate x . fmap readName
+intercalateName (Name x) = Name . intercalate x . fmap unName
 {-# INLINE intercalateName #-}
 
 toFieldName :: TypeName -> FieldName
-toFieldName = FieldName . readTypeName
+toFieldName = Name . unName
 {-# INLINE toFieldName #-}
-
--- TypeName
-newtype TypeName = TypeName {readTypeName :: Text}
-  deriving
-    (Generic)
-  deriving newtype
-    ( Show,
-      Ord,
-      Eq,
-      IsString,
-      ToString,
-      Hashable,
-      Semigroup,
-      FromJSON,
-      ToJSON
-    )
-
-instance Lift TypeName where
-  lift = liftString . readTypeName
-
-#if MIN_VERSION_template_haskell(2,16,0)
-  liftTyped = liftTypedString . readTypeName
-#endif
 
 liftTypedString :: IsString a => Token -> Q (TExp a)
 liftTypedString = unsafeTExpCoerce . stringE . T.unpack
@@ -238,12 +192,6 @@ liftTypedString = unsafeTExpCoerce . stringE . T.unpack
 liftString :: Token -> ExpQ
 liftString = stringE . T.unpack
 {-# INLINE liftString #-}
-
-instance Msg TypeName where
-  msg TypeName {readTypeName} = Message $ "\"" <> readTypeName <> "\""
-
-instance RenderGQL TypeName where
-  renderGQL = fromText . readTypeName
 
 -- Description
 type Description = Text
@@ -264,9 +212,13 @@ instance Eq Position where
 
 data GQLError = GQLError
   { message :: Message,
-    locations :: [Position]
+    locations :: [Position],
+    extensions :: Maybe Value
   }
-  deriving (Show, Eq, Generic, FromJSON, ToJSON)
+  deriving (Show, Eq, Generic, FromJSON)
+
+instance ToJSON GQLError where
+  toJSON = genericToJSON (defaultOptions {omitNothingFields = True})
 
 type GQLErrors = [GQLError]
 
@@ -371,9 +323,9 @@ isReserved _ = False
 {-# INLINE isReserved #-}
 
 convertToJSONName :: FieldName -> FieldName
-convertToJSONName (FieldName hsName)
-  | not (T.null hsName) && isReserved (FieldName name) && (T.last hsName == '\'') = FieldName name
-  | otherwise = FieldName hsName
+convertToJSONName (Name hsName)
+  | not (T.null hsName) && isReserved (Name name) && (T.last hsName == '\'') = Name name
+  | otherwise = Name hsName
   where
     name = T.init hsName
 {-# INLINE convertToJSONName #-}